split out array stream
