+++ python/src/nanoarrow/array_stream.py
@@ -21,7 +21,7 @@ from typing import Iterable, Tuple
 from nanoarrow._lib import CMaterializedArrayStream
 from nanoarrow._repr_utils import make_class_label
 from nanoarrow.array import Array
-from nanoarrow.c_lib import c_array_stream
+from nanoarrow.c_array_stream import c_array_stream
 from nanoarrow.iterator import iter_py, iter_tuples
 from nanoarrow.schema import Schema
